Add files via upload
The file summarises the inventory management system of a bookstore. 

The tab Vendor Info is given, it shows the ID of books, the contact information of the vendor, and the lead time and quantity of a reorder.

How it works:
After deliveries, in the tab Inventory, you insert the inventory level of the books currently for sale in the bookstore and enter any books (that are offered in the Venfor Info list) expected to be in demand. For the books that are not currently for sale, you need to enter the "cost per unit", "sale price per unit" and "reorder threshold." 

Results:
Reorder status and ETA are automatically updated. These reorders will be summerized in the tab Order Summary, where you can find the reorder quantity, cost, contact details